# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: voice
    description: Text-to-speech voice synthesis
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /voice/speak:
    post:
      operationId: voiceSpeak
      tags: [voice]
      summary: Synthesize speech via ElevenLabs
      description: |
        Synthesizes the given text into speech using the supplied ElevenLabs
        voice id. Responses are cached on disk by the server keyed on the
        sha256 of the (voiceId, settings, text) tuple, so identical requests
        are served from cache without re-billing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceSpeakRequest"
      responses:
        "200":
          description: MP3 audio bytes
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceErrorResponse"
        "502":
          description: Upstream ElevenLabs failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    VoiceSpeakRequest:
      type: object
      required:
        - text
        - voiceId
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 4000
          description: The text to speak. Trimmed and bounded to keep the cache key sane.
        voiceId:
          type: string
          minLength: 1
          description: ElevenLabs voice id (a stable id from voiceProfiles.ts).
        modelId:
          type: string
          description: ElevenLabs model id. Defaults to "eleven_multilingual_v2".
        settings:
          $ref: "#/components/schemas/VoiceSettings"
    VoiceSettings:
      type: object
      properties:
        stability:
          type: number
          minimum: 0
          maximum: 1
        similarityBoost:
          type: number
          minimum: 0
          maximum: 1
        style:
          type: number
          minimum: 0
          maximum: 1
        useSpeakerBoost:
          type: boolean
    VoiceErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        detail:
          type: string