# lib/api-zod/src/generated/types/voiceSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VoiceSettings {
  /**
   * @minimum 0
   * @maximum 1
   */
  stability?: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  similarityBoost?: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  style?: number;
  useSpeakerBoost?: boolean;
}