# lib/api-zod/src/generated/types/voiceSpeakRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VoiceSettings } from "./voiceSettings";

export interface VoiceSpeakRequest {
  /**
   * The text to speak. Trimmed and bounded to keep the cache key sane.
   * @minLength 1
   * @maxLength 4000
   */
  text: string;
  /**
   * ElevenLabs voice id (a stable id from voiceProfiles.ts).
   * @minLength 1
   */
  voiceId: string;
  /** ElevenLabs model id. Defaults to "eleven_multilingual_v2". */
  modelId?: string;
  settings?: VoiceSettings;
}